Move IWindowsDriverSession to Interfaces folder and rename WindowsDriverSessionEx.
