+++ tests/IC.Tests.App.Poms/Appium/Interfaces/IFacade.cs
@@ -1,6 +1,7 @@
 ﻿using IC.Navigation;
 using IC.Navigation.Extensions.Appium;
 using IC.Navigation.Extensions.Appium.WindowsDriver;
+using IC.Navigation.Extensions.Appium.WindowsDriver.Interfaces;
 using IC.Navigation.Interfaces;
 using OpenQA.Selenium.Appium.Windows;
 using System;
